Broke out the abuf parsing into a helper module
